refactor: sopht_mpi/sopht_simulator -> sopht_mpi/simulator
