+++ sopht_mpi/simulator/immersed_body/cosserat_rod/cosserat_rod_flow_interaction_mpi.py
@@ -1,7 +1,7 @@
 __all__ = ["CosseratRodFlowInteraction"]
 from elastica.rod.cosserat_rod import CosseratRod
 import numpy as np
-from sopht_mpi.sopht_mpi_simulator.immersed_body import (
+from sopht_mpi.simulator.immersed_body import (
     ImmersedBodyForcingGrid,
     EmptyForcingGrid,
     ImmersedBodyFlowInteractionMPI,
