+++ sopht_mpi/simulator/immersed_body/immersed_body_flow_interaction_mpi.py
@@ -1,7 +1,7 @@
 __all__ = ["ImmersedBodyFlowInteractionMPI"]
 import numpy as np
 from sopht_mpi.numeric.immersed_boundary_ops import VirtualBoundaryForcingMPI
-from sopht_mpi.sopht_mpi_simulator.immersed_body.immersed_body_forcing_grid import (
+from sopht_mpi.simulator.immersed_body.immersed_body_forcing_grid import (
     ImmersedBodyForcingGrid,
 )
 from mpi4py import MPI
