+++ sopht_mpi/simulator/immersed_body/rigid_body/rigid_body_flow_interaction_mpi.py
@@ -1,7 +1,7 @@
 __all__ = ["RigidBodyFlowInteractionMPI"]
 from elastica import RigidBodyBase
 import numpy as np
-from sopht_mpi.sopht_mpi_simulator.immersed_body import (
+from sopht_mpi.simulator.immersed_body import (
     ImmersedBodyForcingGrid,
     EmptyForcingGrid,
     ImmersedBodyFlowInteractionMPI,
