--- sopht_mpi/sopht_mpi_simulator/immersed_body/__init__.py
@@ -1,5 +1,0 @@
-from sopht_mpi.sopht_mpi_simulator.immersed_body.immersed_body_flow_interaction_mpi import *
-from sopht_mpi.sopht_mpi_simulator.immersed_body.immersed_body_forcing_grid import *
-from sopht_mpi.sopht_mpi_simulator.immersed_body.rigid_body import *
-from sopht_mpi.sopht_mpi_simulator.immersed_body.cosserat_rod import *
-from sopht_mpi.sopht_mpi_simulator.immersed_body.flow_forces import *
