--- sopht_mpi/sopht_mpi_simulator/immersed_body/cosserat_rod/__init__.py
@@ -1,2 +1,0 @@
-from sopht_mpi.sopht_mpi_simulator.immersed_body.cosserat_rod.cosserat_rod_forcing_grids import *
-from sopht_mpi.sopht_mpi_simulator.immersed_body.cosserat_rod.cosserat_rod_flow_interaction_mpi import *
