--- sopht_mpi/sopht_mpi_simulator/immersed_body/rigid_body/__init__.py
@@ -1,2 +1,0 @@
-from sopht_mpi.sopht_mpi_simulator.immersed_body.rigid_body.rigid_body_forcing_grids import *
-from sopht_mpi.sopht_mpi_simulator.immersed_body.rigid_body.rigid_body_flow_interaction_mpi import *
