+++ tests/test_simulator/immersed_body/rigid_body/test_mpi_rigid_body_flow_interaction.py
@@ -1,6 +1,6 @@
 import pytest
 import numpy as np
-import sopht_mpi.sopht_mpi_simulator as sps
+import sopht_mpi.simulator as sps
 from sopht.utils.precision import get_real_t
 from tests.test_simulator.immersed_body.rigid_body.test_rigid_body_forcing_grids import (
     mock_2d_cylinder,
