+++ tests/test_simulator/immersed_body/rigid_body/test_rigid_body_forcing_grids.py
@@ -1,7 +1,7 @@
 import elastica as ea
 import numpy as np
 import pytest
-import sopht_mpi.sopht_mpi_simulator as sps
+import sopht_mpi.simulator as sps
 from sopht.utils.precision import get_test_tol
