+++ tests/test_simulator/immersed_body/test_flow_forces.py
@@ -1,5 +1,5 @@
 import numpy as np
-import sopht_mpi.sopht_mpi_simulator as sps
+import sopht_mpi.simulator as sps
 
 
 class MockBodyFlowInteractor:
